update SDK test
