--- rust/integration-tests/expected_db_output_files/imported_testnet_txns/events_processor/278556781/events.json
@@ -1,34 +1,0 @@
-[
-  {
-    "sequence_number": 2,
-    "creation_number": 0,
-    "account_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862",
-    "transaction_version": 278556781,
-    "transaction_block_height": 20040326,
-    "type_": "0x1::account::CoinRegisterEvent",
-    "data": {
-      "type_info": {
-        "module_name": "0x70726f746f6e5f63626479",
-        "struct_name": "0x5052545f4b4642434343",
-        "account_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862"
-      }
-    },
-    "inserted_at": "2024-10-11T19:02:56.746418",
-    "event_index": 0,
-    "indexed_type": "0x1::account::CoinRegisterEvent"
-  },
-  {
-    "sequence_number": 0,
-    "creation_number": 10,
-    "account_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862",
-    "transaction_version": 278556781,
-    "transaction_block_height": 20040326,
-    "type_": "0x1::coin::DepositEvent",
-    "data": {
-      "amount": "1000000000000000000"
-    },
-    "inserted_at": "2024-10-11T19:02:56.746418",
-    "event_index": 1,
-    "indexed_type": "0x1::coin::DepositEvent"
-  }
-]
