--- rust/integration-tests/expected_db_output_files/imported_testnet_txns/fungible_asset_processor/1255836496/fungible_asset_activities.json
@@ -1,74 +1,0 @@
-[
-  {
-    "transaction_version": 1255836496,
-    "event_index": -1,
-    "owner_address": "0xf21d4c76354db6492841e31fefeaa5a3e8a1071d2bcbbdd1cf1cbc6fa3954b88",
-    "storage_id": "0x2086a7aaf18928233884d1725d4f92c5f26c67ac03cb98fe2dc40c99671282dc",
-    "asset_type": "0x1::aptos_coin::AptosCoin",
-    "is_frozen": null,
-    "amount": "50500",
-    "type_": "0x1::aptos_coin::GasFeeEvent",
-    "is_gas_fee": true,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0xfabb471223cefd7064b1f19f6e0e06468d2bd830d8ae832916406dd8297098af::meso::deposit_coin",
-    "block_height": 295534162,
-    "token_standard": "v1",
-    "transaction_timestamp": "2024-06-18T05:07:55",
-    "storage_refund_amount": "0"
-  },
-  {
-    "transaction_version": 1255836496,
-    "event_index": 0,
-    "owner_address": "0xf21d4c76354db6492841e31fefeaa5a3e8a1071d2bcbbdd1cf1cbc6fa3954b88",
-    "storage_id": "0x2086a7aaf18928233884d1725d4f92c5f26c67ac03cb98fe2dc40c99671282dc",
-    "asset_type": "0x1::aptos_coin::AptosCoin",
-    "is_frozen": null,
-    "amount": "300000",
-    "type_": "0x1::coin::WithdrawEvent",
-    "is_gas_fee": false,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0xfabb471223cefd7064b1f19f6e0e06468d2bd830d8ae832916406dd8297098af::meso::deposit_coin",
-    "block_height": 295534162,
-    "token_standard": "v1",
-    "transaction_timestamp": "2024-06-18T05:07:55",
-    "storage_refund_amount": "0"
-  },
-  {
-    "transaction_version": 1255836496,
-    "event_index": 1,
-    "owner_address": "0x6cbd58aec3b1eb5aff37ece4dff62b829ecbc6b851ee94056e852bcde5690ab4",
-    "storage_id": "0x56c216a20a151df2e917008e27f3b70250f1cbb80eeef24e64e2335630825fa3",
-    "asset_type": "0x1::aptos_coin::AptosCoin",
-    "is_frozen": null,
-    "amount": "300000",
-    "type_": "0x1::coin::DepositEvent",
-    "is_gas_fee": false,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0xfabb471223cefd7064b1f19f6e0e06468d2bd830d8ae832916406dd8297098af::meso::deposit_coin",
-    "block_height": 295534162,
-    "token_standard": "v1",
-    "transaction_timestamp": "2024-06-18T05:07:55",
-    "storage_refund_amount": "0"
-  },
-  {
-    "transaction_version": 1255836496,
-    "event_index": 2,
-    "owner_address": "0xbe1390db94c2f8e83fd870c195656921ed4a2a5dee644c6d9c3168fea082a828",
-    "storage_id": "0x8f4f0991e18bd9620a4b3756dbe15afd20135eaaa93e0d0b82e6e2412a8e12b8",
-    "asset_type": "0x4551f2167c446b219c4dee2aa5496395c82a465cd57c5b362621af451c5d062b",
-    "is_frozen": null,
-    "amount": "300000",
-    "type_": "0x1::fungible_asset::Deposit",
-    "is_gas_fee": false,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0xfabb471223cefd7064b1f19f6e0e06468d2bd830d8ae832916406dd8297098af::meso::deposit_coin",
-    "block_height": 295534162,
-    "token_standard": "v2",
-    "transaction_timestamp": "2024-06-18T05:07:55",
-    "storage_refund_amount": "0"
-  }
-]
