--- rust/integration-tests/expected_db_output_files/imported_testnet_txns/fungible_asset_processor/278556781/fungible_asset_activities.json
@@ -1,38 +1,0 @@
-[
-  {
-    "transaction_version": 278556781,
-    "event_index": -1,
-    "owner_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862",
-    "storage_id": "0xa0f5bae99aac34d548f1c5f487800e375e451c1d9576bcdae6f3902752c83c01",
-    "asset_type": "0x1::aptos_coin::AptosCoin",
-    "is_frozen": null,
-    "amount": "2787000",
-    "type_": "0x1::aptos_coin::GasFeeEvent",
-    "is_gas_fee": true,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0x1::code::publish_package_txn",
-    "block_height": 20040326,
-    "token_standard": "v1",
-    "transaction_timestamp": "2022-10-13T01:46:57",
-    "storage_refund_amount": "0"
-  },
-  {
-    "transaction_version": 278556781,
-    "event_index": 1,
-    "owner_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862",
-    "storage_id": "0xfc6edbad85409914ab088b68fed0528e4d8e36dcee31075bb8252b4f426a6cff",
-    "asset_type": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862::proton_cbdy::PRT_KFBCCC",
-    "is_frozen": null,
-    "amount": "1000000000000000000",
-    "type_": "0x1::coin::DepositEvent",
-    "is_gas_fee": false,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0x1::code::publish_package_txn",
-    "block_height": 20040326,
-    "token_standard": "v1",
-    "transaction_timestamp": "2022-10-13T01:46:57",
-    "storage_refund_amount": "0"
-  }
-]
