--- rust/integration-tests/expected_db_output_files/imported_testnet_txns/fungible_asset_processor/5979639459/fungible_asset_activities.json
@@ -1,56 +1,0 @@
-[
-  {
-    "transaction_version": 5979639459,
-    "event_index": -1,
-    "owner_address": "0x54ce130990ecf55cd12de74bff03dcfaacf63066e0945d3c415d61644303da39",
-    "storage_id": "0x32edf0b123cd57094ba7b0b5a0b1524e0f2117b547adf65e8b0c08b39e30d242",
-    "asset_type": "0x1::aptos_coin::AptosCoin",
-    "is_frozen": null,
-    "amount": "99900",
-    "type_": "0x1::aptos_coin::GasFeeEvent",
-    "is_gas_fee": true,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0x1::aptos_account::transfer",
-    "block_height": 354619634,
-    "token_standard": "v1",
-    "transaction_timestamp": "2024-09-13T16:48:29",
-    "storage_refund_amount": "0"
-  },
-  {
-    "transaction_version": 5979639459,
-    "event_index": 1,
-    "owner_address": "0x54ce130990ecf55cd12de74bff03dcfaacf63066e0945d3c415d61644303da39",
-    "storage_id": "0x32edf0b123cd57094ba7b0b5a0b1524e0f2117b547adf65e8b0c08b39e30d242",
-    "asset_type": "0x1::aptos_coin::AptosCoin",
-    "is_frozen": null,
-    "amount": "10",
-    "type_": "0x1::coin::WithdrawEvent",
-    "is_gas_fee": false,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0x1::aptos_account::transfer",
-    "block_height": 354619634,
-    "token_standard": "v1",
-    "transaction_timestamp": "2024-09-13T16:48:29",
-    "storage_refund_amount": "0"
-  },
-  {
-    "transaction_version": 5979639459,
-    "event_index": 2,
-    "owner_address": "0xb7a4a81a3d513e3e18ee6bec61a001d7e18c5d92bb6645a2f21f0b0fec2531a6",
-    "storage_id": "0x239f9719aceb8223231f615e436fcfc8b7e7a6cb03350d00dc72b47678e5d78b",
-    "asset_type": "0x1::aptos_coin::AptosCoin",
-    "is_frozen": null,
-    "amount": "10",
-    "type_": "0x1::coin::DepositEvent",
-    "is_gas_fee": false,
-    "gas_fee_payer_address": null,
-    "is_transaction_success": true,
-    "entry_function_id_str": "0x1::aptos_account::transfer",
-    "block_height": 354619634,
-    "token_standard": "v1",
-    "transaction_timestamp": "2024-09-13T16:48:29",
-    "storage_refund_amount": "0"
-  }
-]
