+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/1/fungible_asset_activities.json
