+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/278556781/current_fungible_asset_balances.json
@@ -1,0 +1,26 @@
+[
+  {
+    "storage_id": "0xa0f5bae99aac34d548f1c5f487800e375e451c1d9576bcdae6f3902752c83c01",
+    "owner_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862",
+    "asset_type": "0x1::aptos_coin::AptosCoin",
+    "is_primary": true,
+    "is_frozen": false,
+    "amount": "10041540182",
+    "last_transaction_timestamp": "2022-10-13T01:46:57",
+    "last_transaction_version": 278556781,
+    "token_standard": "v1",
+    "inserted_at": "2024-10-15T05:16:56.910478"
+  },
+  {
+    "storage_id": "0xfc6edbad85409914ab088b68fed0528e4d8e36dcee31075bb8252b4f426a6cff",
+    "owner_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862",
+    "asset_type": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862::proton_cbdy::PRT_KFBCCC",
+    "is_primary": true,
+    "is_frozen": false,
+    "amount": "1000000000000000000",
+    "last_transaction_timestamp": "2022-10-13T01:46:57",
+    "last_transaction_version": 278556781,
+    "token_standard": "v1",
+    "inserted_at": "2024-10-15T05:16:56.910478"
+  }
+]
