+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/278556781/fungible_asset_metadata.json
@@ -1,0 +1,20 @@
+[
+  {
+    "asset_type": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862::proton_cbdy::PRT_KFBCCC",
+    "creator_address": "0x589dabf3b43ef5c331ea3c0f9b6b005860c1f0ef94a5948d133e051b20c22862",
+    "name": "Aptos",
+    "symbol": "APT",
+    "decimals": 10,
+    "icon_uri": null,
+    "project_uri": null,
+    "last_transaction_version": 278556781,
+    "last_transaction_timestamp": "2022-10-13T01:46:57",
+    "supply_aggregator_table_handle_v1": null,
+    "supply_aggregator_table_key_v1": null,
+    "token_standard": "v1",
+    "is_token_v2": null,
+    "inserted_at": "2024-10-15T05:16:56.771727",
+    "supply_v2": null,
+    "maximum_v2": null
+  }
+]
