+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/5523474016/fungible_asset_balances.json
@@ -1,0 +1,1 @@
+[]
