+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/5979639459/coin_supply.json
@@ -1,0 +1,11 @@
+[
+  {
+    "transaction_version": 5979639459,
+    "coin_type_hash": "91ceb1308a98389691e05158b07ed5f079ab78461a6bb8d5a4054b1bb5cb8bb6",
+    "coin_type": "0x1::aptos_coin::AptosCoin",
+    "supply": "29164832318786122251",
+    "transaction_timestamp": "2024-09-13T16:48:29",
+    "transaction_epoch": 17778,
+    "inserted_at": "2024-10-15T05:16:54.402430"
+  }
+]
