+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/5979639459/current_fungible_asset_balances.json
@@ -1,0 +1,26 @@
+[
+  {
+    "storage_id": "0x239f9719aceb8223231f615e436fcfc8b7e7a6cb03350d00dc72b47678e5d78b",
+    "owner_address": "0xb7a4a81a3d513e3e18ee6bec61a001d7e18c5d92bb6645a2f21f0b0fec2531a6",
+    "asset_type": "0x1::aptos_coin::AptosCoin",
+    "is_primary": true,
+    "is_frozen": false,
+    "amount": "10",
+    "last_transaction_timestamp": "2024-09-13T16:48:29",
+    "last_transaction_version": 5979639459,
+    "token_standard": "v1",
+    "inserted_at": "2024-10-15T05:16:54.392129"
+  },
+  {
+    "storage_id": "0x32edf0b123cd57094ba7b0b5a0b1524e0f2117b547adf65e8b0c08b39e30d242",
+    "owner_address": "0x54ce130990ecf55cd12de74bff03dcfaacf63066e0945d3c415d61644303da39",
+    "asset_type": "0x1::aptos_coin::AptosCoin",
+    "is_primary": true,
+    "is_frozen": false,
+    "amount": "16673484350",
+    "last_transaction_timestamp": "2024-09-13T16:48:29",
+    "last_transaction_version": 5979639459,
+    "token_standard": "v1",
+    "inserted_at": "2024-10-15T05:16:54.392129"
+  }
+]
