+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/5992795934/current_fungible_asset_balances.json
@@ -1,0 +1,38 @@
+[
+  {
+    "storage_id": "0x5120dc6832daa352228ba75c1e7e73f57a80b608c900a7060382516a5c035f20",
+    "owner_address": "0xd34ea8c6faa9d25a9f07504759c1d92bd5ad05af89062d57cf69dd4861fe3e3c",
+    "asset_type": "0x1::aptos_coin::AptosCoin",
+    "is_primary": true,
+    "is_frozen": false,
+    "amount": "18901942640",
+    "last_transaction_timestamp": "2024-09-16T06:58:10",
+    "last_transaction_version": 5992795934,
+    "token_standard": "v1",
+    "inserted_at": "2024-10-15T05:16:56.081056"
+  },
+  {
+    "storage_id": "0x8bd5745daa2d6e5817819c6bf37600196e55cfba8c5bcdfdf55684bf4accbb9c",
+    "owner_address": "0x422e4585fdb8705aeaafb9501e7e03dc8125134da04ed8322244687c50c3fa55",
+    "asset_type": "0x618d700859af35c08518d1552c2601808261288cfd032461fb3424587484ac7f",
+    "is_primary": true,
+    "is_frozen": false,
+    "amount": "0",
+    "last_transaction_timestamp": "2024-09-16T06:58:10",
+    "last_transaction_version": 5992795934,
+    "token_standard": "v2",
+    "inserted_at": "2024-10-15T05:16:56.081056"
+  },
+  {
+    "storage_id": "0xfe90bdd996728b93e09f5d59fd91d6d44be85ab7acea5d1ed535b4c13b2c32c0",
+    "owner_address": "0x13fc6cb932ab90e65ce901460ce8d4b232d4e365294c6ea5afca94625adba47f",
+    "asset_type": "0x618d700859af35c08518d1552c2601808261288cfd032461fb3424587484ac7f",
+    "is_primary": true,
+    "is_frozen": false,
+    "amount": "9996985969300000000",
+    "last_transaction_timestamp": "2024-09-16T06:58:10",
+    "last_transaction_version": 5992795934,
+    "token_standard": "v2",
+    "inserted_at": "2024-10-15T05:16:56.081056"
+  }
+]
