+++ rust/integration-tests/sdk_expected_db_output_files/imported_testnet_txns/fungible_asset_processor/5992795934/fungible_asset_metadata.json
@@ -1,0 +1,1 @@
+[]
