+++ rust/sdk-processor/src/processors/events_processor.rs
@@ -53,83 +53,6 @@ impl EventsProcessor {
             },
         }
     }
-    //
-    // pub async fn run_processor(self) -> Result<()> {
-    //     let processor_name = self.config.processor_config.name();
-    //
-    //     // Run migrations
-    //     match self.config.db_config {
-    //         DbConfig::PostgresConfig(ref postgres_config) => {
-    //             run_migrations(
-    //                 postgres_config.connection_string.clone(),
-    //                 self.db_pool.clone(),
-    //             )
-    //             .await;
-    //         },
-    //     }
-    //
-    //     //  Merge the starting version from config and the latest processed version from the DB
-    //     let starting_version = get_starting_version(&self.config, self.db_pool.clone()).await?;
-    //
-    //     // Check and update the ledger chain id to ensure we're indexing the correct chain
-    //     let grpc_chain_id = TransactionStream::new(self.config.transaction_stream_config.clone())
-    //         .await?
-    //         .get_chain_id()
-    //         .await?;
-    //     check_or_update_chain_id(grpc_chain_id as i64, self.db_pool.clone()).await?;
-    //
-    //     let processor_config = match self.config.processor_config {
-    //         ProcessorConfig::EventsProcessor(processor_config) => processor_config,
-    //         _ => {
-    //             return Err(anyhow::anyhow!(
-    //                 "Invalid processor config for EventsProcessor: {:?}",
-    //                 self.config.processor_config
-    //             ))
-    //         },
-    //     };
-    //     let channel_size = processor_config.channel_size;
-    //
-    //     // Define processor steps
-    //     let transaction_stream = TransactionStreamStep::new(TransactionStreamConfig {
-    //         starting_version: Some(starting_version),
-    //         ..self.config.transaction_stream_config
-    //     })
-    //     .await?;
-    //     let events_extractor = EventsExtractor {};
-    //     let events_storer = EventsStorer::new(self.db_pool.clone(), processor_config);
-    //     let order_step = OrderByVersionStep::new(
-    //         starting_version,
-    //         Duration::from_secs(UPDATE_PROCESSOR_STATUS_SECS),
-    //     );
-    //     let version_tracker =
-    //         LatestVersionProcessedTracker::new(self.db_pool.clone(), processor_name.to_string());
-    //
-    //     // Connect processor steps together
-    //     let (_, buffer_receiver) = ProcessorBuilder::new_with_inputless_first_step(
-    //         transaction_stream.into_runnable_step(),
-    //     )
-    //     .connect_to(events_extractor.into_runnable_step(), channel_size)
-    //     .connect_to(events_storer.into_runnable_step(), channel_size)
-    //     .connect_to(order_step.into_runnable_step(), channel_size)
-    //     .connect_to(version_tracker.into_runnable_step(), channel_size)
-    //     .end_and_return_output_receiver(channel_size);
-    //
-    //     // (Optional) Parse the results
-    //     loop {
-    //         match buffer_receiver.recv().await {
-    //             Ok(txn_context) => {
-    //                 debug!(
-    //                     "Finished processing events from versions [{:?}, {:?}]",
-    //                     txn_context.metadata.start_version, txn_context.metadata.end_version,
-    //                 );
-    //             },
-    //             Err(e) => {
-    //                 info!("No more transactions in channel: {:?}", e);
-    //                 break Ok(());
-    //             },
-    //         }
-    //     }
-    // }
 }
 
 #[async_trait::async_trait]
